requirements.txt
# Azure dependencies
azure-search-documents>=11.4.0
azure-core>=1.29.5
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# OpenAI dependencies
openai>=1.3.7
tiktoken>=0.5.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
openpyxl>=3.1.0

# Environment and configuration
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.2.0

# HTTP client
httpx>=0.25.2

# JSON handling
orjson>=3.9.10

# Date/time handling
python-dateutil>=2.8.2

# MongoDB dependencies
pymongo>=4.6.0
motor>=3.3.0

# Web scraping dependencies
apify-client>=1.7.0